fix: CustomsItem.Quantity should be an integer

+++ ShipEngine/Models/Dto/Common/Customs.cs
@@ -41,7 +41,7 @@ namespace ShipEngineSDK.Common
         /// <summary>
         /// The quantity of this item in the shipment.
         /// </summary>
-        public double? Quantity { get; set; }
+        public int? Quantity { get; set; }
 
         /// <summary>
         /// The declared customs value of each item
